made changes on module.

+++ omod/src/main/java/org/openmrs/module/smsreminder/scheduler/SendSmsReminderTask.java
@@ -1,6 +1,7 @@
 package org.openmrs.module.smsreminder.scheduler;
 
 import java.util.Calendar;
+import java.util.Iterator;
 import java.util.UUID;
 
 import org.openmrs.GlobalProperty;
@@ -14,6 +15,7 @@ import org.openmrs.module.smsreminder.model.MessageSent;
 import org.openmrs.module.smsreminder.model.MessageToBeSent;
 import org.openmrs.module.smsreminder.model.NotificationPatient;
 import org.openmrs.module.smsreminder.utils.DatasUtil;
+import org.openmrs.module.smsreminder.utils.Validator;
 import org.openmrs.module.smsreminder.webservice.Consumer;
 import org.openmrs.scheduler.tasks.AbstractTask;
 
@@ -24,7 +26,7 @@ public class SendSmsReminderTask extends AbstractTask {
 	final AdministrationService administrationService = Context.getAdministrationService();
 	final GlobalProperty gpUs = administrationService.getGlobalPropertyObject("smsreminder.location_id");
 	final GlobalProperty prefix = administrationService.getGlobalPropertyObject("smsreminder.prefix");
-	final GlobalProperty partial_message = administrationService.getGlobalPropertyObject("smsreminder.partial_message");
+	final GlobalProperty partialMessage = administrationService.getGlobalPropertyObject("smsreminder.partial_message");
 	final SmsReminderService smsReminderService = SmsReminderUtils.getService();
 	final PatientService patientService = Context.getPatientService();
 	final LocationService locationService = Context.getLocationService();
@@ -36,7 +38,7 @@ public class SendSmsReminderTask extends AbstractTask {
 
 			for (NotificationPatient notificationPatient : smsReminderService.getAllNotificationPatient()) {
 				String mensage = "Sr(a) ".concat(notificationPatient.getFullName())
-						.concat(partial_message.getPropertyValue())
+						.concat(partialMessage.getPropertyValue())
 						.concat(locationService.getLocation(Integer.valueOf(gpUs.getPropertyValue())).getName())
 						.concat(" no dia ").concat(DatasUtil.formatarDataPt(notificationPatient.getNextVisitDate()));
 				createMessageToBeSent(notificationPatient, mensage);
@@ -50,25 +52,33 @@ public class SendSmsReminderTask extends AbstractTask {
 				String partnerMsgId = UUID.randomUUID().toString();
 
 				try {
-					ScheduleResult result = Consumer.sendMensage(messageToBeSent.getMessage(),
-							prefix.getPropertyValue() + messageToBeSent.getPhoneNumber(), partnerMsgId);
+					String[] splitNumber = messageToBeSent.getPhoneNumber().split(",");
 
-					MessageSent mensageSent = new MessageSent();
-					mensageSent.setNid(messageToBeSent.getNid());
-					mensageSent.setFullName(messageToBeSent.getFullName());
-					mensageSent.setLastVisitDate(messageToBeSent.getLastVisitDate());
-					mensageSent.setNextVisitDate(messageToBeSent.getNextVisitDate());
-					mensageSent.setMessage(messageToBeSent.getMessage());
-					mensageSent.setPartnerMsgId(partnerMsgId);
-					mensageSent.setPhoneNumber(messageToBeSent.getPhoneNumber());
-					mensageSent.setPatient(patientService.getPatient(messageToBeSent.getPatientId()));
-					mensageSent.setDateCreated(Calendar.getInstance().getTime());
-					mensageSent.setReminderDays(messageToBeSent.getReminderDays());
-					mensageSent.setGender(messageToBeSent.getGender());
+					for (int i = 0; i < splitNumber.length; i++) {
 
-					if (result.getScheduleStatus().name().equals("SCHEDULE_OK")) {
-						smsReminderService.saveMensageSent(mensageSent);
-						smsReminderService.deleteMessageToBeSent(messageToBeSent);
+						if (Validator.cellNumberValidator(splitNumber[i])) {
+
+							ScheduleResult result = Consumer.sendMensage(messageToBeSent.getMessage(),
+									prefix.getPropertyValue() + splitNumber[i], partnerMsgId);
+
+							MessageSent mensageSent = new MessageSent();
+							mensageSent.setNid(messageToBeSent.getNid());
+							mensageSent.setFullName(messageToBeSent.getFullName());
+							mensageSent.setLastVisitDate(messageToBeSent.getLastVisitDate());
+							mensageSent.setNextVisitDate(messageToBeSent.getNextVisitDate());
+							mensageSent.setMessage(messageToBeSent.getMessage());
+							mensageSent.setPartnerMsgId(partnerMsgId);
+							mensageSent.setPhoneNumber(messageToBeSent.getPhoneNumber());
+							mensageSent.setPatient(patientService.getPatient(messageToBeSent.getPatientId()));
+							mensageSent.setDateCreated(Calendar.getInstance().getTime());
+							mensageSent.setReminderDays(messageToBeSent.getReminderDays());
+							mensageSent.setGender(messageToBeSent.getGender());
+
+							if (result.getScheduleStatus().name().equals("SCHEDULE_OK")) {
+								smsReminderService.saveMensageSent(mensageSent);
+								smsReminderService.deleteMessageToBeSent(messageToBeSent);
+							}
+						}
 					}
 				} catch (Throwable e) {
 					e.printStackTrace();
